+++ lib/aws/lex/conversation/type/dialog_action.rb
@@ -8,10 +8,12 @@ module Aws
           include Base
 
           optional :slot_to_elicit
+          optional :slot_elicitation_style, default: -> { 'Default' }
           required :type
 
           coerce(
-            type: DialogActionType
+            type: DialogActionType,
+            slot_elicitation_style: SlotElicitationStyle
           )
         end
       end